Fix some issues when I moved over the TM code
Started seeing some losses on time, but in a rather odd way.
Once there was a single loss on time, almost all games were.
Noticed some mistakes when I moved over the TM code to time.c
If I still notice them, I will have to revert back...

SPRT @ 1.0+0.01s Threads=1 Hash=2MB
LLR: 2.97 (-2.94, 2.94) [-3.00, 1.00]
Games: 8125 W: 2712 L: 2542 D: 2871

SPRT @ 5.0+0.05s Threads=1 Hash=2MB
LLR: 1.04 (-2.94, 2.94) [-3.00, 1.00]
Games: 2225 W: 603 L: 547 D: 1075

BENCH : 9,962,753

+++ src/time.h
@@ -25,8 +25,8 @@ typedef struct Manager {
     
     // History of Iterative Deepening
     int values[MAX_DEPTH];
-    int bestMoves[MAX_DEPTH];
-    int timeUsage[MAX_DEPTH];
+    uint16_t bestMoves[MAX_DEPTH];
+    double timeUsage[MAX_DEPTH];
     int depth;
     
     // Base time of the search
